requirements.txt
quart>=0.19.6
hypercorn>=0.17.3
boto3>=1.28.0
twilio>=8.0.0
google-generativeai>=0.3.0
soundfile>=0.12.0
websockets>=11.0.0
requests>=2.31.0
amazon-transcribe>=0.6.0
azure-cognitiveservices-speech>=1.3.0
